+++ app/pkg/logsviewer/eval_viewer.go
@@ -3,9 +3,10 @@ package logsviewer
 import (
 	"context"
 	"fmt"
-	parserv1 "github.com/stateful/runme/v3/pkg/api/gen/proto/go/runme/parser/v1"
 	"net/http"
 	"strings"
+
+	parserv1 "github.com/stateful/runme/v3/pkg/api/gen/proto/go/runme/parser/v1"
 
 	"connectrpc.com/connect"
 	"github.com/go-logr/zapr"
